peer manager redefined
