+++ src/Nethermind/Nethermind.Network/IPeerManager.cs
@@ -20,11 +20,10 @@ using System.Threading.Tasks;
 
 namespace Nethermind.Network
 {
-    public interface IPeerManager : IPeerStatusProvider
+    public interface IPeerManager
     {
-        void Initialize(bool isDiscoveryEnabled);
+        Task Init(bool isDiscoveryEnabled);
         Task Start();
         Task StopAsync();
-        Task RunPeerUpdate();
     }
 }
