+++ src/Nethermind/Nethermind.Network/PeerManager.cs
@@ -20,9 +20,18 @@ using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
+using System.Net;
 using System.Text;
 using System.Threading.Tasks;
 using System.Timers;
+using DotNetty.Buffers;
+using DotNetty.Codecs;
+using DotNetty.Common.Internal.Logging;
+using DotNetty.Handlers.Logging;
+using DotNetty.Transport.Bootstrapping;
+using DotNetty.Transport.Channels;
+using DotNetty.Transport.Channels.Sockets;
+using Microsoft.Extensions.Logging.Console;
 using Nethermind.Blockchain;
 using Nethermind.Config;
 using Nethermind.Core;
@@ -37,6 +46,7 @@ using Nethermind.Network.Discovery.RoutingTable;
 using Nethermind.Network.P2P;
 using Nethermind.Network.P2P.Subprotocols.Eth;
 using Nethermind.Network.Rlpx;
+using Nethermind.Network.Rlpx.Handshake;
 using Nethermind.Network.Stats;
 
 namespace Nethermind.Network
@@ -45,8 +55,15 @@ namespace Nethermind.Network
     /// </summary>
     public class PeerManager : IPeerManager
     {
-        private readonly IRlpxPeer _localPeer;
+        private readonly int _localPort;
+        private readonly IEncryptionHandshakeService _encryptionHandshakeService;
+        private readonly IMessageSerializationService _serializationService;
+        private IChannel _bootstrapChannel;
+        private IEventLoopGroup _bossGroup;
+        private IEventLoopGroup _workerGroup;
+        
         private readonly ILogger _logger;
+        private readonly ILogManager _logManager;
         private readonly IDiscoveryManager _discoveryManager;
         private readonly INetworkConfig _networkConfig;
         private readonly ISynchronizationManager _synchronizationManager;
@@ -67,7 +84,11 @@ namespace Nethermind.Network
         private readonly ConcurrentDictionary<NodeId, Peer> _activePeers = new ConcurrentDictionary<NodeId, Peer>();
         private readonly ConcurrentDictionary<NodeId, Peer> _candidatePeers = new ConcurrentDictionary<NodeId, Peer>();
 
-        public PeerManager(IRlpxPeer localPeer,
+        public PeerManager(
+            NodeId localNodeId, // TODO: move to network config
+            int localPort, // TODO: move to network config
+            IEncryptionHandshakeService encryptionHandshakeService,
+            IMessageSerializationService messageSerializationService,
             IDiscoveryManager discoveryManager,
             ISynchronizationManager synchronizationManager,
             INodeStatsProvider nodeStatsProvider,
@@ -77,8 +98,8 @@ namespace Nethermind.Network
             IPerfService perfService,
             ILogManager logManager)
         {
-            _logger = logManager?.GetClassLogger() ?? throw new ArgumentNullException(nameof(logManager));
-            _localPeer = localPeer;
+            _logManager = logManager ?? throw new ArgumentNullException(nameof(logManager));
+            _logger = logManager.GetClassLogger();
             _networkConfig = configurationProvider.GetConfig<INetworkConfig>();
             _synchronizationManager = synchronizationManager;
             _nodeStatsProvider = nodeStatsProvider;
@@ -87,22 +108,20 @@ namespace Nethermind.Network
             _nodeFactory = nodeFactory;
             _peerStorage = peerStorage;
             _peerStorage.StartBatch();
+            
+            _nodeStatsProvider = nodeStatsProvider;
+            _logger = logManager.GetClassLogger();
+
+            _synchronizationManager = synchronizationManager ?? throw new ArgumentNullException(nameof(synchronizationManager));
+            _encryptionHandshakeService = encryptionHandshakeService ?? throw new ArgumentNullException(nameof(encryptionHandshakeService));
+            _serializationService = messageSerializationService ?? throw new ArgumentNullException(nameof(messageSerializationService));
+
+            LocalNodeId =  localNodeId ?? throw new ArgumentNullException(nameof(localNodeId));
+            _localPort = localPort;
         }
 
         internal IReadOnlyCollection<Peer> CandidatePeers => _candidatePeers.Values.ToArray();
         internal IReadOnlyCollection<Peer> ActivePeers => _activePeers.Values.ToArray();
-
-        public void Initialize(bool isDiscoveryEnabled)
-        {
-            _isDiscoveryEnabled = isDiscoveryEnabled;
-            _discoveryManager.NodeDiscovered += async (s, e) => await OnNodeDiscovered(s, e);
-            _localPeer.OutConnectionInitialized += OnRemoteOutConnectionInitialized;
-            _localPeer.HandshakeInitialized += async (s, e) => await OnHandshakeInitialized(s, e);
-            _synchronizationManager.SyncEvent += async (s, e) => await OnSyncEvent(s, e);
-
-            LoadConfiguredtrustedPeers();
-            LoadPeersFromDb();
-        }
 
         public async Task Start()
         {
@@ -245,11 +264,6 @@ namespace Nethermind.Network
             _isPeerUpdateInProgress = false;
         }
 
-        public bool IsPeerConnected(NodeId peerId)
-        {
-            return _activePeers.ContainsKey(peerId);
-        }
-
         private void LogSessionStats()
         {
             if (_logger.IsInfo)
@@ -342,7 +356,7 @@ namespace Nethermind.Network
         {
             try
             {
-                await _localPeer.ConnectAsync(candidate.Node.Id, candidate.Node.Host, candidate.Node.Port, candidate.NodeStats);
+                await ConnectAsync(candidate.Node.Id, candidate.Node.Host, candidate.Node.Port, candidate.NodeStats);
                 return true;
             }
             catch (NetworkingException ex)
@@ -1129,6 +1143,187 @@ namespace Nethermind.Network
             public ISynchronizationPeer SynchronizationPeer { get; set; }
             public IP2PMessageSender P2PMessageSender { get; set; }
             public ClientConnectionType ClientConnectionType { get; set; }
+        }
+        
+        public NodeId LocalNodeId { get; }
+
+        public async Task Shutdown()
+        {
+//            InternalLoggerFactory.DefaultFactory.AddProvider(new ConsoleLoggerProvider((s, level) => true, false));
+
+            await _bootstrapChannel.CloseAsync().ContinueWith(t =>
+            {
+                if (t.IsFaulted)
+                {
+                    _logger.Error($"{nameof(Shutdown)} failed", t.Exception);
+                }
+            });
+
+            await Task.WhenAll(_bossGroup.ShutdownGracefullyAsync(), _workerGroup.ShutdownGracefullyAsync())
+                .ContinueWith(t =>
+                {
+                    if (t.IsFaulted)
+                    {
+                        _logger.Error($"Groups shutdown failed in {nameof(PeerManager)}", t.Exception);
+                    }
+                });
+        }
+
+        public async Task Init(bool isDiscoveryEnabled)
+        {
+            if (_isInitialized)
+            {
+                throw new InvalidOperationException($"{nameof(PeerManager)} already initialized.");
+            }
+
+            _isInitialized = true;
+
+            try
+            {
+                _bossGroup = new MultithreadEventLoopGroup(1);
+                _workerGroup = new MultithreadEventLoopGroup();
+
+                ServerBootstrap bootstrap = new ServerBootstrap();
+                bootstrap
+                    .Group(_bossGroup, _workerGroup)
+                    .Channel<TcpServerSocketChannel>()
+                    .ChildOption(ChannelOption.SoBacklog, 100)
+                    .Handler(new LoggingHandler("BOSS", DotNetty.Handlers.Logging.LogLevel.TRACE))
+                    .ChildHandler(new ActionChannelInitializer<ISocketChannel>(ch =>
+                        InitializeChannel(ch, EncryptionHandshakeRole.Recipient, null,
+                            ((IPEndPoint) ch.RemoteAddress).Address.ToString(), ((IPEndPoint) ch.RemoteAddress).Port)));
+
+                _bootstrapChannel = await bootstrap.BindAsync(_localPort).ContinueWith(t =>
+                {
+                    if (t.IsFaulted)
+                    {
+                        _logger.Error($"{nameof(Init)} failed", t.Exception);
+                        return null;
+                    }
+
+                    return t.Result;
+                });
+
+                if (_bootstrapChannel == null)
+                {
+                    throw new NetworkingException($"Failed to initialize {nameof(_bootstrapChannel)}", NetwokExceptionType.Other);
+                }
+            }
+            catch (Exception ex)
+            {
+                _logger.Error($"{nameof(Init)} failed.", ex);
+                // TODO: check what happens on nulls
+                await Task.WhenAll(_bossGroup?.ShutdownGracefullyAsync(), _workerGroup?.ShutdownGracefullyAsync());
+                throw;
+            }
+            
+            _isDiscoveryEnabled = isDiscoveryEnabled;
+            _discoveryManager.NodeDiscovered += async (s, e) => await OnNodeDiscovered(s, e);
+            OutConnectionInitialized += OnRemoteOutConnectionInitialized;
+            HandshakeInitialized += async (s, e) => await OnHandshakeInitialized(s, e);
+            _synchronizationManager.SyncEvent += async (s, e) => await OnSyncEvent(s, e);
+
+            LoadConfiguredtrustedPeers();
+            LoadPeersFromDb();
+        }
+
+        public async Task ConnectAsync(NodeId remoteId, string host, int port, INodeStats nodeStats)
+        {
+            if (_logger.IsTrace) _logger.Trace($"Connecting to {remoteId}@{host}:{port}");
+
+            Bootstrap clientBootstrap = new Bootstrap();
+            clientBootstrap.Group(_workerGroup);
+            clientBootstrap.Channel<TcpSocketChannel>();
+
+            clientBootstrap.Option(ChannelOption.TcpNodelay, true);
+            clientBootstrap.Option(ChannelOption.MessageSizeEstimator, DefaultMessageSizeEstimator.Default);
+            clientBootstrap.Option(ChannelOption.ConnectTimeout, Timeouts.InitialConnection);
+            clientBootstrap.RemoteAddress(host, port);
+
+            clientBootstrap.Handler(new ActionChannelInitializer<ISocketChannel>(ch => InitializeChannel(ch, EncryptionHandshakeRole.Initiator, remoteId, host, port, nodeStats)));
+
+            var connectTask = clientBootstrap.ConnectAsync(new IPEndPoint(IPAddress.Parse(host), port));
+            var firstTask = await Task.WhenAny(connectTask,
+                Task.Delay(Timeouts.InitialConnection.Add(TimeSpan.FromSeconds(5))));
+            if (firstTask != connectTask)
+            {
+                if (_logger.IsTrace) _logger.Trace($"Connection timed out: {remoteId}@{host}:{port}");
+                throw new NetworkingException($"Failed to connect to {remoteId} (timeout)", NetwokExceptionType.Timeout);
+            }
+
+            if (connectTask.IsFaulted)
+            {
+                if (_logger.IsTrace)
+                {
+                    _logger.Trace($"Error when connecting to {remoteId}@{host}:{port}, error: {connectTask.Exception}");
+                }
+
+                throw new NetworkingException($"Failed to connect to {remoteId}", NetwokExceptionType.TargetUnreachable,connectTask.Exception);
+            }
+
+            if (_logger.IsTrace) _logger.Trace($"Connected to {remoteId}@{host}:{port}");
+        }
+
+        public event EventHandler<ConnectionInitializedEventArgs> OutConnectionInitialized;
+        public event EventHandler<ConnectionInitializedEventArgs> HandshakeInitialized;
+
+        private void InitializeChannel(IChannel channel, EncryptionHandshakeRole role, NodeId remoteId = null, string remoteHost = null, int? remotePort = null, INodeStats nodeStats = null)
+        {
+            var connectionType = role == EncryptionHandshakeRole.Recipient ? ClientConnectionType.In : ClientConnectionType.Out;
+            P2PSession p2PSession = new P2PSession(
+                LocalNodeId,
+                _localPort,
+                _serializationService,
+                _synchronizationManager,
+                _logManager, _nodeStatsProvider, nodeStats)
+            {
+                ClientConnectionType = connectionType
+            };
+
+            //This is the first moment we get confirmed publicKey of remote node in case of outgoing connections
+            if (connectionType == ClientConnectionType.Out)
+            {
+                if (_logger.IsTrace)
+                {
+                    _logger.Trace($"Initializing {connectionType.ToString().ToUpper()} channel{(connectionType == ClientConnectionType.Out ? $": {remoteId}@{remoteHost}:{remotePort}" : string.Empty)}");
+                }
+
+                p2PSession.RemoteNodeId = remoteId;
+                p2PSession.RemoteHost = remoteHost;
+                p2PSession.RemotePort = remotePort;
+
+                OutConnectionInitialized?.Invoke(this, new ConnectionInitializedEventArgs(p2PSession, connectionType));
+            }
+
+            var handshakeHandler = new NettyHandshakeHandler(_encryptionHandshakeService, p2PSession, role, remoteId, _logManager);
+            handshakeHandler.HandshakeInitialized += (s, e) =>
+            {
+                //This is the first moment we get confirmed publicKey of remote node in case of incoming connections
+                if (connectionType == ClientConnectionType.In)
+                {
+                    if (_logger.IsTrace)
+                    {
+                        _logger.Trace($"Handshake initialized {connectionType.ToString().ToUpper()} channel {p2PSession.RemoteNodeId}@{p2PSession.RemoteHost}:{p2PSession.RemotePort}");
+                    }
+                }
+
+                HandshakeInitialized?.Invoke(this, new ConnectionInitializedEventArgs(p2PSession, connectionType));
+            };
+
+            IChannelPipeline pipeline = channel.Pipeline;
+            pipeline.AddLast(new LoggingHandler(connectionType.ToString().ToUpper(), DotNetty.Handlers.Logging.LogLevel.TRACE));
+            pipeline.AddLast("enc-handshake-dec", new LengthFieldBasedFrameDecoder(ByteOrder.BigEndian, ushort.MaxValue, 0, 2, 0, 0, true));
+            pipeline.AddLast("enc-handshake-handler", handshakeHandler);
+
+            channel.CloseCompletion.ContinueWith(async x =>
+            {
+                if (_logger.IsTrace)
+                {
+                    _logger.Trace($"Channel disconnected: {p2PSession.RemoteNodeId}");
+                }
+
+                await p2PSession.DisconnectAsync(DisconnectReason.ClientQuitting, DisconnectType.Remote);
+            });
         }
     }
 }
